requirements.txt
bertopic==0.16.4
fpdf==1.7.2
numpy==1.24.3
pandas==2.2.3
pdfplumber==0.11.4
python-dotenv==1.0.1
python_docx==1.1.2
scikit_learn==1.5.2
sentence_transformers==3.1.0
streamlit==1.39.0
transformers==4.44.2
